refactor: split grafana integration into modular components
