+++ src/controllers/query.controller.js
@@ -1,0 +1,38 @@
+import { models } from '../models/models.js';
+import { getSQLFromSequelize, parseSQLQuery } from '../utils/sqlQueryBuilder.js';
+
+export async function createQuery(req, res) {
+  const { model, operation, options } = req.body;
+  try {
+    const response = await getSQLFromSequelize(
+      models[model],
+      operation,
+      options
+    );
+    console.log('SQL Query:', response);
+    return res.status(200).json({
+      message: 'SQL query created successfully',
+      query: response,
+    });
+  } catch (error) {
+    return res.status(500).json({
+      message: 'Failed to create SQL query',
+      error: error.message,
+    });
+  }
+}
+
+export async function parseQuery(req, res) {
+  try {
+    const response = parseSQLQuery(req.body.rawSql);
+    return res.status(200).json({
+      message: 'SQL query parsed successfully',
+      sql: response,
+    });
+  } catch (error) {
+    return res.status(500).json({
+      message: 'Failed to parse SQL query',
+      error: error.message,
+    });
+  }
+}
